MockWebhook: track shutdown status more gracefully (#4761)
* MockWebhook: track shutdown status more gracefully

When we close the server sometimes the messages are interrupted and
can't be decoded by the webhook. This change accepts any decoding
failure after server shutdown.


* Call NewWebhook passing 't', and protect against nil cmd in Terminate


---------

+++ test/testutils/acceptance.go
@@ -383,7 +383,7 @@ func (amc *AlertmanagerCluster) Terminate() {
 // data.
 func (am *Alertmanager) Terminate() {
 	am.T.Helper()
-	if am.cmd.Process != nil {
+	if am.cmd != nil && am.cmd.Process != nil {
 		if err := syscall.Kill(am.cmd.Process.Pid, syscall.SIGTERM); err != nil {
 			am.T.Logf("Error sending SIGTERM to Alertmanager process: %v", err)
 		}
